+++ passport.js
@@ -2,30 +2,48 @@ import GitHubStrategy from 'passport-github'
 import passport from 'passport'
 import dotenv from 'dotenv'
 dotenv.config();
+import {app} from './app.js'
 
 
-app.use(passport.initialize());
 
-GitHubStrategy.Strategy;
-
+//client check by codes
 passport.use(new GitHubStrategy({
-    clientID: process.env.GITHUB_CLIENT_ID,
+    clientID: process.env.GITHUB_ID,
     clientSecret: process.env.GITHUB_SECRET,
     callbackURL: "http://localhost:3000/auth/github/callback"
   },
   function(accessToken, refreshToken, profile, cb) {
     console.log(profile);
-    callback(null, profile);
+    cb(null, profile);
   }
 ));
 
 
-//authorisation
-app.get('/auth/github', passport.authenticate('github', { scope: [ 'user:email' ] }));
+//getting id from user profile
+function extractingID(){
+    app.use(passport.initialize());
+    app.use(passport.session());
+    passport.serializeUser(function (user,cb) {
+      cb(null, user.id)
+    });
+    passport.deserializeUser(function(id,cb) {
+      cb(null, id)
+    });  
+}
 
-app.get('/auth/github/callback', 
-  passport.authenticate('github', { failureRedirect: '/login' }),
-  function(req, res) {
-    // Successful authentication, redirect home.
-    res.redirect('/');
-  });
+//authorisation
+function authorisation(){
+    app.get('/auth/github', passport.authenticate('github'));
+    app.get(
+        '/auth/github/callback',
+         passport.authenticate('github', { failureRedirect: '/login' }),
+         function(req, res){
+        // Successful authentication, redirect home.
+        res.redirect('/');
+       }
+    );
+}
+
+
+
+export {passport, extractingID, authorisation};
