+++ protection.js
@@ -1,0 +1,10 @@
+//protection
+function isAuth (req, res, next){
+    if(req.user){
+        next();
+    }else{
+        res.redirect('/login');
+    }
+}
+
+export {isAuth};
